passados salva_rastro_solicitacao() para inicio de fluxo.
