+++ sme_terceirizadas/dados_comuns/behaviors.py
@@ -217,76 +217,3 @@ class Logs(object):
     @property
     def logs(self):
         return LogSolicitacoesUsuario.objects.filter(uuid_original=self.uuid)
-
-
-class TemRastroInstituicaoPartindoDaEscola(models.Model):
-    rastro_escola = models.ForeignKey('escola.Escola',
-                                      on_delete=models.DO_NOTHING,
-                                      null=True,
-                                      blank=True,
-                                      related_name='%(app_label)s_%(class)s_rastro_escola',
-                                      editable=False)
-    rastro_dre = models.ForeignKey('escola.DiretoriaRegional',
-                                   on_delete=models.DO_NOTHING,
-                                   null=True,
-                                   related_name='%(app_label)s_%(class)s_rastro_dre',
-                                   blank=True,
-                                   editable=False)
-    rastro_lote = models.ForeignKey('escola.Lote',
-                                    on_delete=models.DO_NOTHING,
-                                    null=True,
-                                    blank=True,
-                                    related_name='%(app_label)s_%(class)s_rastro_lote',
-                                    editable=False)
-    rastro_terceirizada = models.ForeignKey('terceirizada.Terceirizada',
-                                            on_delete=models.DO_NOTHING,
-                                            null=True,
-                                            blank=True,
-                                            related_name='%(app_label)s_%(class)s_rastro_terceirizada',
-                                            editable=False)
-
-    def salva_rastro_solicitacao(self):
-        self.rastro_escola = self.escola
-        self.rastro_dre = self.escola.diretoria_regional
-        self.rastro_lote = self.escola.lote
-        self.rastro_terceirizada = self.escola.lote.terceirizada
-        self.save()
-
-    class Meta:
-        abstract = True
-
-
-class TemRastroInstituicaoPartindoDaDiretoriaRegional(models.Model):
-    rastro_escolas = models.ManyToManyField('escola.Escola',
-                                            blank=True,
-                                            related_name='%(app_label)s_%(class)s_rastro_escola',
-                                            editable=False)
-    rastro_dre = models.ForeignKey('escola.DiretoriaRegional',
-                                   on_delete=models.DO_NOTHING,
-                                   null=True,
-                                   related_name='%(app_label)s_%(class)s_rastro_dre',
-                                   blank=True,
-                                   editable=False)
-    rastro_lote = models.ForeignKey('escola.Lote',
-                                    on_delete=models.DO_NOTHING,
-                                    null=True,
-                                    blank=True,
-                                    related_name='%(app_label)s_%(class)s_rastro_lote',
-                                    editable=False)
-    rastro_terceirizada = models.ForeignKey('terceirizada.Terceirizada',
-                                            on_delete=models.DO_NOTHING,
-                                            null=True,
-                                            blank=True,
-                                            related_name='%(app_label)s_%(class)s_rastro_terceirizada',
-                                            editable=False)
-
-    def salva_rastro_solicitacao(self):
-        escolas = [i.escola for i in self.escolas_quantidades.all()]
-        self.rastro_escolas.set(escolas)
-        self.rastro_dre = self.diretoria_regional
-        self.rastro_lote = self.lote
-        self.rastro_terceirizada = self.lote.terceirizada
-        self.save()
-
-    class Meta:
-        abstract = True
